requirements.txt
Flask
google-generativeai
termcolor
prettytable
openai